pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "webtest-framework"
version = "1.0.0"
description = "Web application testing framework with Playwright"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "pytest>=7.4.0",
    "pytest-playwright>=0.4.4",
    "playwright>=1.40.0",
    "pytest-html>=4.1.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "smoke: Quick smoke tests",
    "regression: Full regression tests",
    "api: API tests",
]

[tool.setuptools.packages.find]
where = ["src"]
